Validação na alteração de imagem de usuário

+++ src/controller/usuario_imagem.php
@@ -38,20 +38,33 @@ function salvarFotoPerfil($usuarioDaoMysql, $formatosImagemPermitidos){
         }
 
         $usuario->setFotoPerfil($fotoPerfil);
-        $usuarioDaoMysql->atualizar($usuario);
+        $atualizouFoto = $usuarioDaoMysql->atualizar($usuario);
 
-        // session_start();
-        $_SESSION['message'] = (Object) [
-            'type'=>'info',
-            'message' => 'Foto de perfil editada com sucesso!'
-        ];
-
+        if($atualizouFoto == false) {
+            $_SESSION['message'] = (Object) [
+                'type'=>'error',
+                'message' => 'Ocorreu um erro inesperao ao atualizar sua foto do perfil'
+            ];
+            header('Location:http://'.$_SERVER['HTTP_HOST'].'/jobfinder/profile');
+            exit;
+        }
+        
         //move_uploaded_file move o arquivo que foi feito upload pelo formulário e salva no diretório files
         move_uploaded_file($_FILES['foto_perfil']['tmp_name'],$_SERVER['DOCUMENT_ROOT'].'/jobfinder/src/files/'.$fotoPerfil);
         $_SESSION['auth'] = serialize($usuario); //Para que a foto seja atualizada na página é necessário renovar a sessão com a atualização da foto
 
         unlink($_SERVER['DOCUMENT_ROOT'].'/jobfinder/src/files/'.$fotoAntiga);
+
+        $_SESSION['message'] = (Object) [
+            'type'=>'info',
+            'message' => 'Foto de perfil editada com sucesso!'
+        ];
         
+    } else {
+        $_SESSION['message'] = (Object) [
+            'type'=>'error',
+            'message' => 'O arquivo possui um formato inválido!'
+        ];
     }
     header('Location:http://'.$_SERVER['HTTP_HOST'].'/jobfinder/profile');
     exit;
